分类模块开发

+++ jc-club-subject/jc-club-infra/src/main/java/com/jingdianjichi/subject/infra/basic/service/SubjectCategoryService.java
@@ -17,7 +17,7 @@ public interface SubjectCategoryService {
      * @param id 主键
      * @return 实例对象
      */
-    SubjectCategory queryById(Long id);
+    SubjectCategory queryById(Integer id);
 
     /**
      * 查询多条数据
